Moved Audio Transcription test to new format
